requirements.txt
asgiref==3.5.2
autopep8==1.6.0
backports.zoneinfo==0.2.1
Django==4.1.1
django-debug-toolbar==3.6.0
djangorestframework==3.13.1
pycodestyle==2.8.0
pytz==2022.2.1
sqlparse==0.4.2
toml==0.10.2
tzdata==2022.2
